Fixing the int64 problem

+++ wolf/models/device.py
@@ -1,6 +1,6 @@
 import base64
 
-from sqlalchemy import Integer, Binary
+from sqlalchemy import Integer, Binary, BigInteger
 from restfulpy.orm import DeclarativeBase, ModifiedMixin, Field
 
 
@@ -9,7 +9,7 @@ class Device(ModifiedMixin, DeclarativeBase):
 
     id = Field(Integer, primary_key=True, protected=True)
 
-    reference_id = Field(Integer, unique=True, index=True)
+    reference_id = Field(BigInteger, unique=True, index=True)
     secret = Field('secret', Binary(32))
 
     def prepare_for_export(self, column, value):
